# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API LIVE TEST",
        "description": "Test API V1",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/cart/list": {
            "get": {
                "tags": [
                    "site_cart"
                ],
                "summary": "site_cart_list",
                "operationId": "site_cart_list",
                "parameters": [
                    {
                        "name": "skip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 0
                    },
                    {
                        "name": "take",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 10
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuthSite": []
                    }
                ]
            }
        },
        "/api/v1/cart": {
            "post": {
                "tags": [
                    "site_cart"
                ],
                "summary": "site_cart_store",
                "operationId": "site_cart_store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "product_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuthSite": []
                    }
                ]
            }
        },
        "/api/v1/cart/update": {
            "put": {
                "tags": [
                    "site_cart"
                ],
                "summary": "cart_update",
                "operationId": "cart_update",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "product_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Returns when customer is not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": ""
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuthSite": []
                    }
                ]
            }
        },
        "/api/v1/cart/{cart_product_id}": {
            "delete": {
                "tags": [
                    "site_cart"
                ],
                "summary": "site_cart_delete",
                "operationId": "site_cart_delete",
                "parameters": [
                    {
                        "name": "cart_product_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuthSite": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Admin",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "Passport",
                "scheme": "bearer"
            },
            "bearerAuthSite": {
                "type": "http",
                "description": "Customer",
                "name": "AuthorizationSite",
                "in": "header",
                "bearerFormat": "Passport",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}